reorganize all files into src folder
