+++ src/domain/services/notification.service.ts
@@ -1,4 +1,3 @@
-// services/notificationService.ts
 import nodemailer from 'nodemailer';
 
 const transporter = nodemailer.createTransport({
