+++ src/infrastructure/routes/user.routes.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { registerUser, loginUser } from '../controllers/userController';
+import { registerUser, loginUser } from '../controllers/user.controller';
 
 export const userRouter = Router();
